extract JS from sidebar.inc.php
