+++ src/js/main.js
@@ -1,6 +1,7 @@
 import * as Ajax from './ajax.js';
 import * as Base from './base.js';
 import * as Search from './search.js';
+import * as Sidebar from './sidebar.js';
 import * as Slideshow from './slideshow.js';
 import * as Tools from './tools.js';
 
@@ -8,4 +9,5 @@ Object.assign(window, Ajax);
 Object.assign(window, Base);
 Object.assign(window, Tools);
 Object.assign(window, Search);
+Object.assign(window, Sidebar);
 Object.assign(window, Slideshow);
