+++ src/js/sidebar.js
@@ -28,3 +28,45 @@ $('#sidebar-header').click(function(){
 
     Cookies.set('sidebar_state', newstate, {jsCookieString});
 });
+
+$(function() {
+    $(".header").click(function () {
+        var $header = $(this);
+        // getting the next element
+        var $content = $header.next();
+        // open up the content needed - toggle the slide- if visible, slide up, if not slidedown.
+        $content.slideToggle(500, function() {
+            $header.children(".header-img").toggleClass("expanded collapsed");
+            var sbstate = "expanded";
+            if ($header.children(".header-img").hasClass("collapsed")) {
+                sbstate = "collapsed";
+            }
+            Cookies.set('sb_' + $header.children(".header-img").attr('id'), sbstate, {jsCookieString});
+        });
+    });
+});
+
+$(document).ready(function() {
+    // Get a string of all the cookies.
+    var cookieArray = document.cookie.split(";");
+    var result = new Array();
+    // Create a key/value array with the individual cookies.
+    for (var elem in cookieArray) {
+        var temp = cookieArray[elem].split("=");
+        // We need to trim whitespaces.
+        temp[0] = $.trim(temp[0]);
+        temp[1] = $.trim(temp[1]);
+        // Only take sb_* cookies (= sidebar cookies)
+        if (temp[0].substring(0, 3) == "sb_") {
+            result[temp[0].substring(3)] = temp[1];
+        }
+    }
+    // Finds the elements and if the cookie is collapsed, it collapsed the found element.
+    for (var key in result) {
+        if ($("#" + key).length && result[key] == "collapsed") {
+            $("#" + key).removeClass("expanded");
+            $("#" + key).addClass("collapsed");
+            $("#" + key).parent().next().slideToggle(0);
+        }
+    }
+});
